first stages of overhauling how debug options and ImGui work together
There is now a new singleton, debugOptions, that holds a whole bunch
of options related to debugging. Code everywhere can refer to these
options if it so desires. ImGui refers to these for all its
operations.

On the plus side, this makes options much more extensible. On the
downside, the core+mappers are now coupled with this system of
debug options. I'm sure that there is some ifdef magic I could
do to make that coupling at least optional, but its nonetheless
not ideal for the long term. In the short term, however, its a
powerful tool.